Port the remaining push button tests
