+++ client/src/test/suite/extension.test.ts
@@ -3,19 +3,15 @@ import fs from 'fs-extra'
 import path from 'path'
 import vscode from 'vscode'
 import SinonRoot from 'sinon'
-import { GitErrorCodes, Repository, CommitOptions, RepositoryState, Branch, RefType } from '../../git-api/git.d'
 import 'source-map-support/register'
 import { expect as expectOrig, getRootPathUri } from './../../utils'
 import { forwardOnDidChangeWorkspaceFolders } from './../../extension'
 import { ImageManagerPanel } from './../../panel-image-manager'
 import { CnxmlPreviewPanel, rawTextHtml, tagElementsWithLineNumbers } from './../../panel-cnxml-preview'
-import * as pushContent from '../../push-content'
-import * as utils from '../../utils'
 import { Suite } from 'mocha'
 import { DOMParser, XMLSerializer } from 'xmldom'
-import { Substitute } from '@fluffy-spoon/substitute'
 import { LanguageClient } from 'vscode-languageclient/node'
-import { EMPTY_BOOKS_AND_ORPHANS, DiagnosticSource, ExtensionServerRequest } from '../../../../common/src/requests'
+import { EMPTY_BOOKS_AND_ORPHANS } from '../../../../common/src/requests'
 import { PanelStateMessageType } from '../../../../common/src/webview-constants'
 import { Disposer, ExtensionEvents, ExtensionHostContext, Panel } from '../../panel'
 
@@ -401,43 +397,6 @@ suite('Extension Test Suite', function (this: Suite) {
     const panel = new CnxmlPreviewPanel({ bookTocs: EMPTY_BOOKS_AND_ORPHANS, resourceRootDir, client: createMockClient(), events: createMockEvents().events })
     await assert.rejects(panel.handleMessage({ type: 'bad-type' } as any))
   })
-  test('canPush returns correct values', async () => {
-    const fileUri = { path: '/test.cnxml', scheme: 'file' } as any as vscode.Uri
-    const cnxmlError = {
-      severity: vscode.DiagnosticSeverity.Error,
-      source: DiagnosticSource.cnxml
-    } as any as vscode.Diagnostic
-    const xmlError = {
-      severity: vscode.DiagnosticSeverity.Error,
-      source: DiagnosticSource.xml
-    } as any as vscode.Diagnostic
-    const errorsBySource = new Map<string, Array<[vscode.Uri, vscode.Diagnostic]>>()
-    const showErrorMsgStub = sinon.stub(vscode.window, 'showErrorMessage')
-
-    // No errors
-    assert(await pushContent.canPush(errorsBySource))
-
-    // CNXML errors
-    errorsBySource.set(DiagnosticSource.cnxml, [[fileUri, cnxmlError]])
-    assert(!(await pushContent.canPush(errorsBySource)))
-    assert(showErrorMsgStub.calledOnceWith(pushContent.PushValidationModal.cnxmlErrorMsg, { modal: true }))
-
-    // Both CNXML and XML errors
-    errorsBySource.clear()
-    showErrorMsgStub.reset()
-    errorsBySource.set(DiagnosticSource.cnxml, [[fileUri, cnxmlError]])
-    errorsBySource.set(DiagnosticSource.xml, [[fileUri, xmlError]])
-    assert(!(await pushContent.canPush(errorsBySource)))
-    assert(showErrorMsgStub.calledOnceWith(pushContent.PushValidationModal.cnxmlErrorMsg, { modal: true }))
-
-    // XML errors, user cancels
-    errorsBySource.clear()
-    showErrorMsgStub.reset()
-    showErrorMsgStub.returns(Promise.resolve(undefined))
-    errorsBySource.set(DiagnosticSource.xml, [[fileUri, xmlError]])
-    assert(!(await pushContent.canPush(errorsBySource)))
-    assert(showErrorMsgStub.calledOnceWith(pushContent.PushValidationModal.xmlErrorMsg, { modal: true }))
-  })
   test('forwardOnDidChangeWorkspaceFolders simply forwards any argument to client', async () => {
     const mockClient = createMockClient()
     const forwarder = forwardOnDidChangeWorkspaceFolders(mockClient)
@@ -516,406 +475,5 @@ suite('Disposables', function (this: Suite) {
     })
     panel.registerDisposable(testDisposable)
     assert(await childDisposed)
-  })
-})
-
-// Push Content Tests
-const ignore = async (message: string): Promise<string | undefined> => { return undefined }
-
-const makeCaptureMessage = (messages: string[]): (message: string) => Promise<string | undefined> => {
-  return async (message: string): Promise<string | undefined> => {
-    messages.push(message)
-    return undefined
-  }
-}
-
-const makeMockDialog = (message: string): () => Promise<string | undefined> => {
-  return async (): Promise<string | undefined> => { return message }
-}
-
-const commitOptions: CommitOptions = { all: true }
-
-export const makeMockNewTag = (tag: string | undefined): (repo: Repository, release: boolean) => string | undefined => {
-  return (): string | undefined => {
-    return tag
-  }
-}
-
-suite('Push Button Test Suite', function (this: Suite) {
-  const sinon = SinonRoot.createSandbox()
-  this.afterEach(() => sinon.restore())
-  const sendRequestMock = sinon.stub()
-  const mockHostContext: ExtensionHostContext = {
-    client: {
-      sendRequest: sendRequestMock
-    }
-  } as any as ExtensionHostContext
-
-  test('getRepo returns repository', async () => {
-    const repo = pushContent.getRepo()
-    assert.notStrictEqual(repo.rootUri, undefined)
-  })
-  test('push with no conflict', async () => {
-    const messages: string[] = []
-    const captureMessage = makeCaptureMessage(messages)
-    const mockMessageInput = makeMockDialog('poet commit')
-
-    const getRepo = (): Repository => {
-      const stubRepo = Substitute.for<Repository>()
-
-      stubRepo.commit('poet commit', commitOptions).resolves()
-      stubRepo.pull().resolves()
-      stubRepo.push().resolves()
-
-      return stubRepo
-    }
-
-    await assert.doesNotReject(pushContent._pushContent(
-      getRepo,
-      mockMessageInput,
-      captureMessage,
-      ignore
-    )())
-    assert.strictEqual(messages.length, 1)
-    assert.strictEqual(messages[0], 'Successful content push.')
-  })
-  test('push with merge conflict', async () => {
-    const messages: string[] = []
-    const captureMessage = makeCaptureMessage(messages)
-    const mockMessageInput = makeMockDialog('poet commit')
-    const error: any = { _fake: 'FakeSoStackTraceIsNotInConsole', message: '' }
-
-    error.gitErrorCode = GitErrorCodes.Conflict
-
-    const getRepo = (): Repository => {
-      const stubRepo = Substitute.for<Repository>()
-
-      stubRepo.commit('poet commit', commitOptions).resolves()
-      stubRepo.pull().rejects(error)
-      stubRepo.push().resolves()
-
-      return stubRepo
-    }
-
-    await assert.doesNotReject(pushContent._pushContent(
-      getRepo,
-      mockMessageInput,
-      ignore,
-      captureMessage
-    )())
-    assert.strictEqual(messages.length, 1)
-    assert.strictEqual(messages[0], 'Content conflict, please resolve.')
-  })
-  test('unknown commit error', async () => {
-    const messages: string[] = []
-    const captureMessage = makeCaptureMessage(messages)
-    const mockMessageInput = makeMockDialog('poet commit')
-    const error: any = { _fake: 'FakeSoStackTraceIsNotInConsole', message: '' }
-
-    error.gitErrorCode = ''
-
-    const getRepo = (): Repository => {
-      const stubRepo = Substitute.for<Repository>()
-
-      stubRepo.commit('poet commit', commitOptions).resolves()
-      stubRepo.pull().rejects(error)
-      stubRepo.push().resolves()
-
-      return stubRepo
-    }
-
-    await assert.doesNotReject(pushContent._pushContent(
-      getRepo,
-      mockMessageInput,
-      ignore,
-      captureMessage
-    )())
-    assert.strictEqual(messages.length, 1)
-    assert.strictEqual(messages[0], 'Push failed: ')
-  })
-  test('push with no changes', async () => {
-    const messages: string[] = []
-    const captureMessage = makeCaptureMessage(messages)
-    const mockMessageInput = makeMockDialog('poet commit')
-    const error: any = { _fake: 'FakeSoStackTraceIsNotInConsole', message: '' }
-
-    error.stdout = 'nothing to commit.'
-
-    const getRepo = (): Repository => {
-      const stubRepo = Substitute.for<Repository>()
-      stubRepo.diffWithHEAD().resolves([])
-      stubRepo.commit('poet commit', commitOptions).rejects(error)
-      stubRepo.pull().resolves()
-      stubRepo.push().resolves()
-
-      return stubRepo
-    }
-
-    await assert.doesNotReject(pushContent._pushContent(
-      getRepo,
-      mockMessageInput,
-      ignore,
-      captureMessage
-    )())
-    assert.strictEqual(messages.length, 1)
-    assert.strictEqual(messages[0], 'No changes to push.')
-  })
-  test('unknown push error', async () => {
-    const messages: string[] = []
-    const captureMessage = makeCaptureMessage(messages)
-    const mockMessageInput = makeMockDialog('poet commit')
-    const error: any = { _fake: 'FakeSoStackTraceIsNotInConsole', message: '' }
-
-    error.stdout = ''
-
-    const getRepo = (): Repository => {
-      const stubRepo = Substitute.for<Repository>()
-
-      stubRepo.commit('poet commit', commitOptions).rejects(error)
-      stubRepo.pull().resolves()
-      stubRepo.push().resolves()
-
-      return stubRepo
-    }
-
-    await assert.doesNotReject(pushContent._pushContent(
-      getRepo,
-      mockMessageInput,
-      ignore,
-      captureMessage
-    )())
-    assert.strictEqual(messages.length, 1)
-    assert.strictEqual(messages[0], 'Push failed: ')
-  })
-  test('pushContent does not invoke _pushContent when canPush is false', async () => {
-    sinon.stub(pushContent, 'openAndValidate').resolves(new Map<string, Array<[vscode.Uri, vscode.Diagnostic]>>())
-    sinon.stub(pushContent, 'canPush').resolves(false)
-    const stubPushContentHelperInner = sinon.stub()
-    sinon.stub(pushContent, '_pushContent').returns(stubPushContentHelperInner)
-    await pushContent.pushContent(mockHostContext)()
-    assert(stubPushContentHelperInner.notCalled)
-    assert(sendRequestMock.notCalled)
-  })
-  test('pushContent invokes _pushContent when canPush is true', async () => {
-    sinon.stub(utils, 'getErrorDiagnosticsBySource').resolves(new Map<string, Array<[vscode.Uri, vscode.Diagnostic]>>())
-    sinon.stub(pushContent, 'getMessage').resolves('poet commit')
-    sinon.stub(pushContent, 'openAndValidate').resolves(new Map<string, Array<[vscode.Uri, vscode.Diagnostic]>>())
-    sinon.stub(pushContent, 'canPush').resolves(true)
-    const stubPushContentHelperInner = sinon.stub()
-    sinon.stub(pushContent, '_pushContent').returns(stubPushContentHelperInner)
-    await pushContent.pushContent(mockHostContext)()
-    assert(stubPushContentHelperInner.calledOnce)
-    assert(sendRequestMock.calledOnceWith(
-      ExtensionServerRequest.BundleEnsureIds
-    ))
-  })
-  test('push to new branch', async () => {
-    const messages: string[] = []
-    const captureMessage = makeCaptureMessage(messages)
-    const mockMessageInput = makeMockDialog('poet commit')
-    const pushStub = sinon.stub()
-    const newBranchName = 'newbranch'
-
-    // This is inconsistent with the rest of this test suite, but it seems we can't use
-    // a Substitute mock for this test case because setting return values on properties
-    // requires disabling strict checking.
-    // (https://github.com/ffMathy/FluffySpoon.JavaScript.Testing.Faking#strict-mode)
-    const getRepo = (): Repository => {
-      const repoBranch = {
-        upstream: undefined,
-        name: newBranchName
-      } as any as Branch
-      const repoState = {
-        HEAD: repoBranch
-      } as any as RepositoryState
-      const stubRepo = {
-        state: repoState,
-        pull: sinon.stub(),
-        push: pushStub,
-        commit: sinon.stub()
-      } as any as Repository
-
-      return stubRepo
-    }
-    await assert.doesNotReject(pushContent._pushContent(
-      getRepo,
-      mockMessageInput,
-      captureMessage,
-      ignore
-    )())
-    assert.strictEqual(messages.length, 1)
-    assert.strictEqual(messages[0], 'Successful content push.')
-    assert(pushStub.calledOnceWith('origin', newBranchName, true))
-  })
-  test('get message returns showInputBox input', async () => {
-    sinon.stub(vscode.window, 'showInputBox').resolves('test')
-    assert.strictEqual(await pushContent.getMessage(), 'test')
-  })
-  test('validateMessage returns "Too short!" for message that is not long enough', async () => {
-    assert.strictEqual(pushContent.validateMessage('a'), 'Too short!')
-  })
-  test('validateMessage returns null for message that is long enough', async () => {
-    assert.strictEqual(pushContent.validateMessage('abc'), null)
-  })
-  test('taggingDialog', async () => {
-    const mockDialog = sinon.stub(vscode.window, 'showInformationMessage')
-    mockDialog.resolves(undefined)
-    assert.strictEqual(await pushContent.taggingDialog(), undefined)
-    mockDialog.resolves(pushContent.Tag.release as any as vscode.MessageItem)
-    assert.strictEqual(await pushContent.taggingDialog(), pushContent.Tag.release)
-    mockDialog.resolves(pushContent.Tag.candidate as any as vscode.MessageItem)
-    assert.strictEqual(await pushContent.taggingDialog(), pushContent.Tag.candidate)
-  })
-  test('getNewTag', async () => {
-    const repoState = {
-      refs: [{
-        name: 'main',
-        type: RefType.Head,
-        commit: 'a'
-      }]
-    } as any as RepositoryState
-    const mockRepo = {
-      state: repoState
-    } as any as Repository
-    const mockHead = {
-      commit: 'a'
-    } as any as Branch
-
-    const showErrorMsgStub = sinon.stub(vscode.window, 'showErrorMessage')
-
-    assert.strictEqual(await pushContent.getNewTag(mockRepo, pushContent.Tag.candidate, mockHead), '1rc')
-    mockRepo.state.refs.push({
-      name: '1rc',
-      type: RefType.Tag,
-      commit: 'b'
-    })
-
-    assert.strictEqual(await pushContent.getNewTag(mockRepo, pushContent.Tag.candidate, mockHead), '2rc')
-    mockRepo.state.refs.push({
-      name: '2rc',
-      type: RefType.Tag,
-      commit: 'a'
-    })
-    assert.strictEqual(await pushContent.getNewTag(mockRepo, pushContent.Tag.candidate, mockHead), undefined)
-    assert(showErrorMsgStub.calledOnceWith('Tag of this type already exists for this content version.', { modal: false }))
-    showErrorMsgStub.reset()
-
-    mockRepo.state.refs.length = 0
-    mockRepo.state.refs.push({
-      name: 'main',
-      type: RefType.Head,
-      commit: 'a'
-    })
-
-    assert.strictEqual(await pushContent.getNewTag(mockRepo, pushContent.Tag.release, mockHead), '1')
-    mockRepo.state.refs.push({
-      name: '1',
-      type: RefType.Tag,
-      commit: 'b'
-    })
-
-    assert.strictEqual(await pushContent.getNewTag(mockRepo, pushContent.Tag.release, mockHead), '2')
-    mockRepo.state.refs.push({
-      name: '2',
-      type: RefType.Tag,
-      commit: 'a'
-    })
-    assert.strictEqual(await pushContent.getNewTag(mockRepo, pushContent.Tag.release, mockHead), undefined)
-    assert(showErrorMsgStub.calledOnceWith('Tag of this type already exists for this content version.', { modal: false }))
-  })
-  test('tagContent', async () => {
-    const showInfoMsgStub = sinon.stub(vscode.window, 'showInformationMessage')
-    const showErrorMsgStub = sinon.stub(vscode.window, 'showErrorMessage')
-    const taggingDialogStub = sinon.stub(pushContent, 'taggingDialog')
-    const getNewTagStub = sinon.stub(pushContent, 'getNewTag')
-    const diffWithHEADStub = sinon.stub()
-    const tagStub = sinon.stub()
-    const pushStub = sinon.stub()
-    const fetchStub = sinon.stub()
-
-    const repoBranch = {
-      name: 'main'
-    } as any as Branch
-    const repoState = {
-      HEAD: repoBranch
-    } as any as RepositoryState
-    const stubRepo = {
-      state: repoState,
-      fetch: fetchStub,
-      diffWithHEAD: diffWithHEADStub,
-      push: pushStub,
-      tag: tagStub
-    } as any as Repository
-
-    sinon.stub(pushContent, 'getRepo').returns(stubRepo)
-
-    // test for dirty workspace
-    diffWithHEADStub.resolves([{}])
-    await pushContent.tagContent()
-    assert(fetchStub.calledOnce)
-    assert(showErrorMsgStub.calledOnceWith('Can\'t tag. Local unpushed changes exist', { modal: false }))
-    fetchStub.reset()
-    showErrorMsgStub.reset()
-
-    // test for canceled tagging
-    taggingDialogStub.resolves(undefined)
-    await pushContent.tagContent()
-    assert(fetchStub.calledOnce)
-    assert(getNewTagStub.notCalled)
-    assert(tagStub.notCalled)
-    fetchStub.reset()
-
-    // test for existing tag
-    diffWithHEADStub.resolves([])
-    taggingDialogStub.resolves(pushContent.Tag.candidate)
-    getNewTagStub.resolves(undefined)
-    await pushContent.tagContent()
-    assert(fetchStub.calledOnce)
-    assert(getNewTagStub.calledOnce)
-    assert(tagStub.notCalled)
-    fetchStub.reset()
-    getNewTagStub.reset()
-
-    // test for valid tag
-    taggingDialogStub.resolves(pushContent.Tag.candidate)
-    getNewTagStub.resolves('1rc')
-    await pushContent.tagContent()
-    assert(fetchStub.calledOnce)
-    assert(getNewTagStub.calledOnce)
-    assert(tagStub.calledOnce)
-    assert(pushStub.calledOnce)
-    assert(showInfoMsgStub.calledOnceWith('Successful tag for Release Candidate.', { modal: false }))
-    fetchStub.reset()
-    getNewTagStub.reset()
-    tagStub.reset()
-    pushStub.reset()
-    showInfoMsgStub.reset()
-    showErrorMsgStub.reset()
-
-    // test for unknown tag error message
-    tagStub.throws()
-    taggingDialogStub.resolves(pushContent.Tag.candidate)
-    getNewTagStub.resolves('1rc')
-    await pushContent.tagContent()
-    assert(showErrorMsgStub.calledOnceWith('Tagging failed: Error', { modal: false }))
-    fetchStub.reset()
-    getNewTagStub.reset()
-    tagStub.reset()
-    pushStub.reset()
-    showInfoMsgStub.reset()
-    showErrorMsgStub.reset()
-
-    // test for unknown push error message
-    tagStub.resolves()
-    pushStub.throws()
-    taggingDialogStub.resolves(pushContent.Tag.candidate)
-    getNewTagStub.resolves('1rc')
-    await pushContent.tagContent()
-    assert(showErrorMsgStub.calledOnceWith('Push failed: Error', { modal: false }))
-    fetchStub.reset()
-    getNewTagStub.reset()
-    tagStub.reset()
-    pushStub.reset()
-    showInfoMsgStub.reset()
   })
 })
